+++ connection_scan_algorithm/include/calculator.hpp
@@ -41,6 +41,7 @@ namespace TrRouting
   class TransitData;
   class ConnectionSet;
   class Point;
+  class GeoFilter;
 
   class Calculator {
 
@@ -48,7 +49,7 @@ namespace TrRouting
 
     std::map<std::string, int> benchmarking;
 
-    Calculator(const TransitData &_transitData);
+    Calculator(const TransitData &_transitData, GeoFilter &_geofilter);
 
     void reset(CommonParameters &parameters, std::optional<std::reference_wrapper<const Point>> origin, std::optional<std::reference_wrapper<const Point>> destination, bool resetAccessPaths = true, bool resetFilters = true);
     // TODO This function supports both allNodes and simple calculation, which
@@ -101,6 +102,8 @@ namespace TrRouting
 
     //TODO set it mutable so it can be changed/reset?
     const TransitData &transitData;
+    //TODO Should it be const?
+    GeoFilter &geoFilter;
 
     int              departureTimeSeconds;
     int              arrivalTimeSeconds;
