+++ tests/connection_scan_algorithm/csa_route_transfer_alternatives_test.cpp
@@ -19,7 +19,6 @@
 #include "trip.hpp"
 #include "od_trip.hpp"
 #include "node.hpp"
-#include "osrm_fetcher.hpp"
 
 /**
  * This file covers tests with transfers and requests for alternatives
@@ -355,7 +354,6 @@ TrRouting::AlternativesResult SingleTAndACalculationFixtureTests::calculateWithA
 {
     // TODO: This needs to be called to set some default values that are still part of the global parameters
     calculator.params.setDefaultValues();
-    TrRouting::OsrmFetcher::birdDistanceAccessibilityEnabled = true;
 
     // TODO Shouldn't need to do this, but we do for now, benchmark needs to be started
     calculator.algorithmCalculationTime.start();
